backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.34.0,<1.0.0
pydantic>=2.0.0,<3.0.0
firebase-admin>=6.0.0,<7.0.0
google-cloud-aiplatform>=1.38.0,<2.0.0
google-cloud-logging>=3.8.0,<4.0.0
python-dotenv>=1.0.0,<2.0.0
